# backend/modules/barberbooking/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API สำหรับโมดูล barberbooking",
        "title": "BarberBooking API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1/barberbooking",
    "paths": {
...
        "/none": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "คืนค่า average rating (float64) ของช่างตัดผมที่ระบุ ภายใน Tenant สำหรับ consistency ของ URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "ดึงคะแนนเฉลี่ยของช่างตัดผม",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "รหัส Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "รหัส Barber",
                        "name": "barber_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "คืนค่า status success และ average rating ใน key `data`",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "number"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid tenant_id หรือ barber_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "ลบ Appointment Review ตามรหัสที่ระบุ โดยระบุ ActorCustomerID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "ลบรีวิว",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "รหัส Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "รหัส Appointment",
                        "name": "appointment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "รหัส Review",
                        "name": "review_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload สำหรับลบรีวิว (actor_customer_id)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/barberBookingController.DeleteReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "คืนค่า status success และข้อความยืนยันการลบ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters หรือ invalid JSON body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "not authorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Review not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/tenants/:tenant_id/appointments/:appointment_id/reviews": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Review"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "รหัส Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "รหัส Appointment",
                        "name": "appointment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload สำหรับสร้างรีวิว",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/barberBookingPort.CreateAppointmentReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "คืนค่า status success และข้อมูล review ที่สร้าง",
                        "schema": {
                            "$ref": "#/definitions/barberBookingModels.AppointmentReview"
                        }
                    },
                    "400": {
                        "description": "Invalid input หรือ cannot review appointment",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tenants/:tenant_id/appointments/:appointment_id/reviews/:review_id": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "แก้ไข Rating และ Comment ของ Appointment Review ตามรหัสที่ระบุ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "อัปเดตรีวิว",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "รหัส Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "รหัส Review",
                        "name": "review_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload สำหรับอัปเดตรีวิว",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/barberBookingPort.UpdateAppointmentReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "คืนค่า status success และข้อมูลรีวิวที่อัปเดต",
                        "schema": {
                            "$ref": "#/definitions/barberBookingModels.AppointmentReview"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters หรือ invalid JSON body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Review not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/tenants/:tenant_id/reviews/:review_id": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "คืนข้อมูล Appointment Review ตามรหัสที่ระบุ ภายใต้ Tenant สำหรับ consistency ของ URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "ดึงรีวิวตาม ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "รหัส Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "รหัส Review",
                        "name": "review_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "คืนค่า status success และข้อมูล review",
                        "schema": {
                            "$ref": "#/definitions/barberBookingModels.AppointmentReview"
                        }
                    },
                    "400": {
                        "description": "Invalid tenant_id หรือ review_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Review not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch review",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "barberBookingController.DeleteReviewRequest": {
            "type": "object",
            "properties": {
                "actor_customer_id": {
                    "type": "integer"
                }
            }
        },
...
        "barberBookingModels.Appointment": {
            "type": "object",
            "properties": {
                "barber": {
                    "$ref": "#/definitions/barberBookingModels.Barber"
                },
                "barber_id": {
                    "type": "integer"
                },
                "branch_id": {
                    "description": "ไม่ preload branch",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/barberBookingModels.Customer"
                },
                "customer_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/barberBookingModels.Service"
                },
                "service_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/barberBookingModels.AppointmentStatus"
                },
                "tenant_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "description": "อ้าง user ที่สร้างคิว (ไม่มี FK)",
                    "type": "integer"
                }
            }
        },
        "barberBookingModels.AppointmentReview": {
            "type": "object",
            "properties": {
                "appointment": {
                    "$ref": "#/definitions/barberBookingModels.Appointment"
                },
                "appointment_id": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/barberBookingModels.Customer"
                },
                "customer_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "barberBookingModels.Barber": {
            "type": "object",
            "properties": {
                "branch_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "tenantID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "barberBookingModels.Customer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "description": "Composite Index",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "description": "optional",
                    "type": "string"
                },
                "tenantID": {
                    "description": "Composite Index",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "barberBookingPort.CreateAppointmentReviewRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string",
                    "example": "Great service!"
                },
                "customer_id": {
                    "type": "integer",
                    "example": 4
                },
                "rating": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "barberBookingPort.UpdateAppointmentReviewRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string",
                    "example": "Updated comment"
                },
                "rating": {
                    "type": "integer",
                    "example": 4
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}